+++ src/RoyalCode.SmartSelector.Demo/Entities/Store/Contact.cs
@@ -1,0 +1,6 @@
+namespace RoyalCode.SmartSelector.Demo.Entities.Store;
+
+public class Contact : Entity<Guid>
+{
+    public required string Email { get; set; }
+}