+++ src/RoyalCode.SmartSelector.Demo/Entities/Store/Supplier.cs
@@ -1,0 +1,10 @@
+namespace RoyalCode.SmartSelector.Demo.Entities.Store;
+
+public class Supplier : Entity<Guid>
+{
+    public required string Name { get; set; }
+
+    public Warehouse? Warehouse { get; set; }
+
+    public ICollection<Contact>? Contacts { get; set; }
+}