+++ src/RoyalCode.SmartSelector.Demo/Entities/Store/Warehouse.cs
@@ -1,0 +1,6 @@
+namespace RoyalCode.SmartSelector.Demo.Entities.Store;
+
+public class Warehouse : Entity<Guid>
+{
+    public required string Location { get; set; }
+}